Update SupportGUI.java
Cleaned up the code.
Added explanatory comments.
Refactored panel names.
